* size_t不是long long型号，混用会导致类型溢出，！=号判断失败。修复了http://demo.lanrenzhijia.com/2015/upload0907/上传 崩溃的问题

+++ net/BlobResourceLoader.cpp
@@ -460,7 +460,7 @@ void BlobResourceLoader::getSizeForNext()
         return;
     }
 
-    size_t size = 0;
+    long long size = 0;
     String filePath;
     blink::WebBlobData::Item* item = m_blobData->items().at(m_sizeItemCount);
     switch (item->type) {
